+++ src/main/java/com/worldcretornica/plotme_core/PlotMe_Core.java
@@ -23,7 +23,7 @@ public class PlotMe_Core {
     private final AbstractSchematicUtil schematicutil;
     private HashMap<String, IPlotMe_GeneratorManager> managers;
     private IWorld worldcurrentlyprocessingexpired;
-    private short counterExpired;
+    private int counterExpired;
     //Spool stuff
     private ConcurrentLinkedQueue<PlotToClear> plotsToClear;
     //Global variables
@@ -229,11 +229,11 @@ public class PlotMe_Core {
         this.worldcurrentlyprocessingexpired = worldcurrentlyprocessingexpired;
     }
 
-    public short getCounterExpired() {
+    public int getCounterExpired() {
         return counterExpired;
     }
 
-    public void setCounterExpired(short counterExpired) {
+    public void setCounterExpired(int counterExpired) {
         this.counterExpired = counterExpired;
     }
